+++ src/ruby/core/drawing.rb
@@ -1,15 +1,3 @@
-# Allows you to call rendering functions within the block but limits the output.
-# to within the bounds of `section`.
-# @yield The block that calls your rendering logic.
-# @param section [Rectangle]
-# @return [nil]
-def scissor_mode(section, &block)
-  begin_scissor_mode(section.x, section.y, section.width, section.height)
-  block.call
-ensure
-  end_scissor_mode
-end
-
 # Allows you to call rendering functions within the block using blending.
 # @yield The block that calls your rendering logic.
 # @param mode [Integer] Must be one of {BLEND_ALPHA}, {BLEND_ADDITIVE},